Added additional progress information
Now shows the total number of iterations at each high water mark. Also
shows the size of the edge pair list associated with each location and
the current index into it. Gives an idea of progress.

Also total number of possible combinations is calculated by just
multiplying the list sizes together. Gives an idea of the size of the
search space at this position on the board that its currently traversed
to.